requirements.txt
amqp==5.1.1
async-timeout==4.0.2      
billiard==3.6.4.0
celery==5.2.6
click==8.1.3
click-didyoumean==0.3.0   
click-plugins==1.1.1      
click-repl==0.2.0
Deprecated==1.2.13        
Flask==2.1.2
Flask-Dropzone==1.6.0     
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
packaging==21.3
pdfminer3k==1.3.4
ply==3.11
prompt-toolkit==3.0.29
pyparsing==3.0.9
PyPDF2==1.28.2
python-dotenv==0.20.0
pytz==2022.1
PyX==0.15
PyYAML==6.0
redis==4.3.1
six==1.16.0
slate3k==0.5.3
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.0